+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Changelog
 - **Fix:** Record IC lookups to `@Multibinds` declarations.
 - **Fix:** Write `@Multibinds` information to metro metadata.
 - **Fix:** Always write metro metadata to `@BindingContainer` classes, even if empty.
+- Add a `ViewModel` assisted injection example to `compose-navigation-app` sample.
+- Small improvements to the doc site (404 page, favicon, etc.)
 
 0.5.2
 -----